PROTO13-afs-attrs-cast-date-to-int-to-avoid-negative-dates-20020516
cheap hack to avoid negative date problem which occasionally plagues solaris


(cherry picked from commit c28c22ceadfb1db8151ae74dacee2a49c42624db)

+++ src/afs/VNOPS/afs_vnop_attrs.c
@@ -95,7 +95,7 @@ int afs_CopyOutAttrs(register struct vcache *avc, register struct vattr *attrs)
     attrs->va_nlink = fakedir ? 100 : avc->m.LinkCount;
     attrs->va_size = fakedir ? 4096 : avc->m.Length;
     attrs->va_atime.tv_sec = attrs->va_mtime.tv_sec = attrs->va_ctime.tv_sec =
-	fakedir ? 0 : avc->m.Date;
+	fakedir ? 0 : (int)avc->m.Date;
     /* set microseconds to be dataversion # so that we approximate NFS-style
      * use of mtime as a dataversion #.  We take it mod 512K because
      * microseconds *must* be less than a million, and 512K is the biggest
